+++ src/pages/QueueStatus/Components/swiper.css
